+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Enhancements
 
 * Add support for sorting `RLMArray`s by multiple columns with `sortedResultsUsingDescriptors:`
+* Add method `deleteAllObjects` on `RLMRealm` to clear a Realm.
 
 ### Bugfixes